BB-190: Remaining forms to ES6 classes
